refactor(production-run): extract the outward projection layer, with the first structural contract
productionRunService.ts 814 -> 658 and leaves the filesize allowlist. The outward projection
layer moves to productionRunProjections.ts: the six Safe*/…Projection types plus
safeRunProjection, runProjection and eventProjection, verbatim. The service keeps
`export type ... from './productionRunProjections'`, so productionRunArtifactOperations.ts and
src/desktop/productionRunBridgeTypes.ts import exactly where they did before.

ServiceDeps, identifier and createProductionRunService stay put on purpose — identifier is
referenced nine times inside the service body and belongs to orchestration, not projection.

This change was blocked earlier today. R21 makes anything under electron/productionRun/ carry a
schema-v3 contract, and that schema only spoke corrective: symptom, direct_cause, class_root,
prevention. A verbatim module move has none of those, so the only ways through were to invent
them or to drop the refactor. It is now the first real user of the `structural` change_kind
added in #425 — the honest description was previously unsayable.

Writing that contract surfaced a real tension worth recording: the structural kind still demands
a CHANGED regression test, and a pure move changes none. The right answer is not to relax that —
it is that a structural extraction should ship the guard that pins what it just moved. So
productionRunProjections.test.ts asserts the extracted surface exists AND compares function
identity between the two modules, which fails if anyone later grows a second copy of the
projection implementation inside the service (P1). TypeScript alone would not catch that: dropping
the re-export only errors at the call sites, never here.

Verified: npx tsc -p electron/tsconfig.json EXIT=0 (types line up across all 16 projection-type
consumers); productionRunService.test.ts + productionRunProjections.test.ts 10/10;
pnpm run check:root-cause-contracts EXIT=0 with the structural cases passing;
node scripts/check-file-sizes.mjs EXIT=0.
